+++ src/main.ts
@@ -2,6 +2,7 @@ import {getHighlighterCore} from 'shiki/core';
 import javascript from 'shiki/langs/javascript.mjs';
 import css from 'shiki/langs/css.mjs';
 import html from 'shiki/langs/html.mjs';
+import json from 'shiki/langs/json.mjs';
 import loadWasm from 'shiki/wasm';
 import monokai from 'shiki/themes/monokai.mjs';
 import nord from 'shiki/themes/nord.mjs';
@@ -22,6 +23,7 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 			javascript,
 			css,
 			html,
+			json,
 		],
 		themes: [
 			monokai,
@@ -30,6 +32,10 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 		loadWasm,
 	});
 	monaco.languages.register({id: 'wikitext', aliases: ['wiki', 'mediawiki']});
+	monaco.languages.register({id: 'javascript', aliases: ['JavaScript', 'js']});
+	monaco.languages.register({id: 'css', aliases: ['CSS']});
+	monaco.languages.register({id: 'html', aliases: ['HTML', 'htm', 'xhtml']});
+	monaco.languages.register({id: 'json', aliases: ['JSON']});
 	shikiToMonaco(highlighter, monaco);
 	monaco.languages.setLanguageConfiguration('wikitext', config);
 
@@ -55,10 +61,10 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 					wikiparse.setConfig(parserConfig as Config);
 				}
 				const linter = new wikiparse.Linter!(true);
-				let timer: NodeJS.Timeout;
+				let timer: number;
 				model.onDidChangeContent(() => {
 					clearTimeout(timer);
-					timer = setTimeout(() => {
+					timer = window.setTimeout(() => {
 						(async () => {
 							monaco.editor.setModelMarkers(model, 'WikiLint', await linter.monaco(model.getValue()));
 						})();